Fix ActiveRecord i18n_scope
