+++ lib/reform/form/active_record.rb
@@ -7,6 +7,12 @@ module Reform::Form::ActiveRecord
       include Reform::Form::ActiveModel
       include Reform::Form::ORM
       extend ClassMethods
+
+      class << self
+        def i18n_scope
+          :activerecord
+        end
+      end
     end
   end
 
@@ -14,27 +20,6 @@ module Reform::Form::ActiveRecord
     def validates_uniqueness_of(attribute, options={})
       options = options.merge(:attributes => [attribute])
       validates_with(UniquenessValidator, options)
-    end
-
-    def i18n_scope
-      :activerecord
-    end
-
-    def human_attribute_name(*args)
-      if model_class
-        model_class.human_attribute_name(*args)
-      else
-        super
-      end
-    end
-
-    def model_class
-      @model_class ||=
-        if Object.const_defined?(model_name.to_s)
-          Object.const_get(model_name.to_s)
-        else
-          nil
-        end
     end
   end
